🔨 refactor: Commands sub-classes
