--- src/classes/Commands/functions/export.ts
@@ -1,12 +1,0 @@
-import * as botStatus from './status';
-import * as help from './help';
-import messageCommand from './message';
-import * as misc from './misc';
-import * as pricelist from './pricelistManager';
-import * as review from './review';
-import * as utils from './utils';
-import * as options from './options';
-import * as manager from './manager';
-import * as request from './request';
-
-export { botStatus, help, messageCommand, misc, pricelist, review, utils, options, manager, request };
